Manually convert behaviour dtype in loss calculation

+++ ncmcm/bundlenet/tests/test_bundlenet_continuous.py
@@ -6,7 +6,7 @@ from ncmcm.bundlenet.bundlenet import BunDLeNet, train_model
 
 def test_bundlenet_training():
     X = np.random.rand(50, 10)
-    B = np.random.random(size=(50,2)).astype(np.float32)
+    B = np.random.random(size=(50,2))
     X_, B_ = prep_data(X, B, win=3)
     # Deploy BunDLe Net
     latent_dim = 3
@@ -28,7 +28,7 @@ def test_bundlenet_training():
 
 def test_bundlenet_training_pca_init():
     X = np.random.rand(50, 10)
-    B = np.random.random(size=(50,2)).astype(np.float32)
+    B = np.random.random(size=(50,2))
     X_, B_ = prep_data(X, B, win=3)
     # Deploy BunDLe Net
     latent_dim = 3
@@ -51,7 +51,7 @@ def test_bundlenet_training_pca_init():
 
 def test_bundlenet_training_best_of_5_init():
     X = np.random.rand(50, 10)
-    B = np.random.random(size=(50,2)).astype(np.float32)
+    B = np.random.random(size=(50,2))
     X_, B_ = prep_data(X, B, win=3)
     # Deploy BunDLe Net
     latent_dim = 3
@@ -75,7 +75,7 @@ def test_bundlenet_training_best_of_5_init():
 
 def test_bundlenet_training_validation_data():
     X = np.random.rand(50, 10)
-    B = np.random.random(size=(50,2)).astype(np.float32)
+    B = np.random.random(size=(50,2))
     X_, B_ = prep_data(X, B, win=3)
     from ncmcm.bundlenet.utils import timeseries_train_test_split
     X_train, X_test, B_train, B_test = timeseries_train_test_split(X_, B_)
